chore: move tests to a separate files
