+++ clients/openframe-client/src/executor/env.rs
@@ -18,21 +18,5 @@ pub(crate) fn apply_env_vars(cmd: &mut Command, env_vars: &[String]) {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn splits_simple() {
-        assert_eq!(split_env("KEY=value"), Some(("KEY", "value")));
-    }
-
-    #[test]
-    fn keeps_equals_in_value() {
-        assert_eq!(split_env("K=a=b"), Some(("K", "a=b")));
-    }
-
-    #[test]
-    fn rejects_without_equals() {
-        assert_eq!(split_env("NOEQ"), None);
-    }
-}
+#[path = "env_tests.rs"]
+mod tests;
