+++ clients/openframe-client/src/logging/platform.rs
@@ -88,20 +88,6 @@ fn can_write_to_directory(path: &PathBuf) -> bool {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
+#[path = "platform_tests.rs"]
+mod tests;
 
-    #[test]
-    fn test_get_log_directory() {
-        let log_dir = get_log_directory();
-
-        #[cfg(target_os = "windows")]
-        assert!(log_dir.to_string_lossy().contains("OpenFrame\\logs"));
-
-        #[cfg(target_os = "macos")]
-        assert_eq!(log_dir.to_string_lossy(), "/Library/Logs/OpenFrame");
-
-        #[cfg(target_os = "linux")]
-        assert_eq!(log_dir.to_string_lossy(), "/var/log/openframe");
-    }
-}
