+++ clients/openframe-client/src/platform/binary_writer.rs
@@ -83,18 +83,5 @@ pub async fn set_executable_permissions(path: &Path) -> Result<()> {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn aside_path_appends_old_to_full_filename() {
-        assert_eq!(
-            aside_path(Path::new(r"C:\x\agent.exe")),
-            PathBuf::from(r"C:\x\agent.exe.old")
-        );
-        assert_eq!(
-            aside_path(Path::new("/x/agent")),
-            PathBuf::from("/x/agent.old")
-        );
-    }
-}
+#[path = "binary_writer_tests.rs"]
+mod tests;
