+++ clients/openframe-client/src/platform/directories.rs
@@ -11,19 +11,16 @@
 ///
 /// The DirectoryManager struct provides a common API that hides platform-specific
 /// implementation details.
-#[allow(unused_imports)] // Context used by unix-only permission paths
 use anyhow::{Context, Result};
 use directories::BaseDirs;
 use std::fs;
 use std::io;
 #[cfg(unix)]
-#[allow(unused_imports)] // used by linux-only permission paths
 use std::os::unix::fs::PermissionsExt;
 use std::path::{Path, PathBuf};
 #[cfg(unix)]
 use std::process::Command;
-#[allow(unused_imports)] // warn used by unix-only paths
-use tracing::{info, warn};
+use tracing::{error, info, warn};
 
 use super::permissions::{PermissionError, Permissions};
 
@@ -284,12 +281,6 @@ pub struct DirectoryManager {
     user_logs_dir: Option<PathBuf>, // For per-user logs when needed
 }
 
-impl Default for DirectoryManager {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
 impl DirectoryManager {
     /// Creates a new DirectoryManager with default platform-specific paths
     pub fn new() -> Self {
@@ -310,7 +301,7 @@ impl DirectoryManager {
         Self {
             logs_dir,
             app_support_dir,
-            secured_dir,
+            secured_dir: secured_dir,
             user_logs_dir: None,
         }
     }
@@ -530,7 +521,6 @@ impl DirectoryManager {
     }
 
     /// Validates permissions for a specific directory
-    #[allow(unused_variables)] // expected_perms used only by unix permission checks
     fn validate_directory_permissions(
         &self,
         path: &Path,
@@ -568,7 +558,6 @@ impl DirectoryManager {
     }
 
     /// Validates permissions for the secured directory (admin-only access)
-    #[allow(unreachable_code)] // macOS path returns early; the tail applies to other unix targets
     fn validate_secured_directory_permissions(&self, path: &Path) -> Result<(), DirectoryError> {
         if !path.exists() {
             return Err(DirectoryError::ValidationFailed(
@@ -733,7 +722,6 @@ impl DirectoryManager {
         let result = std::fs::OpenOptions::new()
             .write(true)
             .create(true)
-            .truncate(true)
             .open(&temp_file);
 
         // Clean up the test file if it was created
@@ -820,7 +808,7 @@ impl DirectoryManager {
     pub fn find_app_bundle_path(executable_path: &Path) -> Option<PathBuf> {
         executable_path
             .ancestors()
-            .find(|p| p.extension().is_some_and(|ext| ext == "app"))
+            .find(|p| p.extension().map_or(false, |ext| ext == "app"))
             .map(|p| p.to_path_buf())
     }
 
@@ -884,268 +872,5 @@ pub async fn remove_app_bundle_path(executable_path: &Path) -> Result<()> {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use tempfile::tempdir;
-
-    #[test]
-    fn test_directory_creation() {
-        let temp_dir = tempdir().unwrap();
-        let logs_dir = temp_dir.path().join("logs");
-        let app_dir = temp_dir.path().join("app");
-        let secured_dir = temp_dir.path().join("secured");
-
-        let manager = DirectoryManager::with_custom_dirs(
-            logs_dir.clone(),
-            app_dir.clone(),
-            secured_dir.clone(),
-        );
-
-        // Test directory creation
-        assert!(manager.ensure_directories().is_ok());
-        assert!(logs_dir.exists());
-        assert!(app_dir.exists());
-    }
-
-    #[test]
-    #[cfg_attr(
-        not(windows),
-        ignore = "requires root-owned secured dir; not applicable to non-root unix runs"
-    )]
-    fn test_directory_permissions() {
-        let temp_dir = tempdir().unwrap();
-        let logs_dir = temp_dir.path().join("logs");
-        let app_dir = temp_dir.path().join("app");
-        let secured_dir = temp_dir.path().join("secured");
-
-        let manager = DirectoryManager::with_custom_dirs(
-            logs_dir.clone(),
-            app_dir.clone(),
-            secured_dir.clone(),
-        );
-
-        // Create directories first
-        assert!(manager.ensure_directories().is_ok());
-
-        // Test permission validation and fixing
-        assert!(manager.validate_permissions().is_ok());
-
-        // Test user directory creation
-        let user_manager = DirectoryManager::with_user_logs_dir();
-        if let Some(user_logs) = &user_manager.user_logs_dir {
-            assert!(
-                user_logs.to_string_lossy().contains("OpenFrame")
-                    || user_logs.to_string_lossy().contains("openframe")
-            );
-        }
-    }
-
-    #[test]
-    fn test_file_permissions() {
-        let temp_dir = tempdir().unwrap();
-        let logs_dir = temp_dir.path().join("logs");
-        let app_dir = temp_dir.path().join("app");
-        let secured_dir = temp_dir.path().join("secured");
-
-        let manager = DirectoryManager::with_custom_dirs(
-            logs_dir.clone(),
-            app_dir.clone(),
-            secured_dir.clone(),
-        );
-
-        // Create directories first
-        assert!(manager.ensure_directories().is_ok());
-
-        // Create a test file in the logs directory
-        let test_file = logs_dir.join("test.log");
-        fs::write(&test_file, "test").unwrap();
-
-        // Apply file permissions
-        let file_perms = Permissions::file();
-        assert!(file_perms.apply(&test_file).is_ok());
-
-        // Verify file permissions
-        #[cfg(unix)]
-        {
-            if unsafe { libc::geteuid() } == 0 {
-                // Only run this check if we're root, otherwise it will fail
-                let metadata = fs::metadata(&test_file).unwrap();
-                assert_eq!(metadata.permissions().mode() & 0o777, 0o644);
-            }
-        }
-    }
-
-    #[test]
-    #[allow(unused_variables)] // manager used only by unix assertion below
-    fn test_error_handling() {
-        // Test with a non-existent directory
-        let non_existent = PathBuf::from("/non_existent_dir_for_test");
-
-        let manager = DirectoryManager::with_custom_dirs(
-            non_existent.clone(),
-            non_existent.clone(),
-            non_existent.clone(),
-        );
-
-        // This should fail on validate because we can't create the directory
-        #[cfg(unix)]
-        if unsafe { libc::geteuid() } != 0 {
-            // We expect this to fail if we're not root
-            assert!(manager.validate_permissions().is_err());
-        }
-    }
-
-    #[test]
-    fn test_user_logs_directory() {
-        let manager = DirectoryManager::with_user_logs_dir();
-
-        // Ensure the user logs directory exists
-        assert!(manager.user_logs_dir.is_some());
-
-        #[cfg(target_os = "macos")]
-        {
-            let user_logs = manager.user_logs_dir.unwrap();
-            assert!(user_logs
-                .to_string_lossy()
-                .contains("Library/Logs/OpenFrame"));
-        }
-
-        #[cfg(target_os = "windows")]
-        {
-            let user_logs = manager.user_logs_dir.unwrap();
-            assert!(user_logs.to_string_lossy().contains("OpenFrame\\Logs"));
-        }
-
-        #[cfg(target_os = "linux")]
-        {
-            let user_logs = manager.user_logs_dir.unwrap();
-            assert!(user_logs
-                .to_string_lossy()
-                .contains(".local/share/openframe/logs"));
-        }
-    }
-
-    #[test]
-    #[cfg_attr(
-        not(windows),
-        ignore = "requires root-owned secured dir; not applicable to non-root unix runs"
-    )]
-    fn test_health_check() {
-        let temp_dir = tempdir().unwrap();
-        let logs_dir = temp_dir.path().join("logs");
-        let app_dir = temp_dir.path().join("app");
-        let secured_dir = temp_dir.path().join("secured");
-
-        let manager = DirectoryManager::with_custom_dirs(
-            logs_dir.clone(),
-            app_dir.clone(),
-            secured_dir.clone(),
-        );
-
-        // Test health check
-        assert!(manager.perform_health_check().is_ok());
-        assert!(logs_dir.exists());
-        assert!(app_dir.exists());
-
-        // Intentionally corrupt permissions to test fixing
-        #[cfg(unix)]
-        {
-            if unsafe { libc::geteuid() } == 0 {
-                // Only run this check if we're root, otherwise it will fail
-                use std::os::unix::fs::PermissionsExt;
-                let bad_perms = fs::Permissions::from_mode(0o700);
-                fs::set_permissions(&logs_dir, bad_perms).unwrap();
-
-                // Health check should fix the permissions
-                assert!(manager.perform_health_check().is_ok());
-
-                // Verify permissions were fixed
-                let metadata = fs::metadata(&logs_dir).unwrap();
-                assert_eq!(metadata.permissions().mode() & 0o777, 0o755);
-            }
-        }
-    }
-
-    #[test]
-    fn test_write_permissions() {
-        let temp_dir = tempdir().unwrap();
-        let logs_dir = temp_dir.path().join("logs");
-        let app_dir = temp_dir.path().join("app");
-        let secured_dir = temp_dir.path().join("secured");
-
-        let manager = DirectoryManager::with_custom_dirs(
-            logs_dir.clone(),
-            app_dir.clone(),
-            secured_dir.clone(),
-        );
-
-        // Create directories first
-        assert!(manager.ensure_directories().is_ok());
-
-        // Test write permissions
-        assert!(manager.can_write_to_directory(&logs_dir));
-        assert!(manager.can_write_to_directory(&app_dir));
-    }
-
-    #[test]
-    fn test_get_logs_directory() {
-        let logs_dir = get_logs_directory();
-
-        #[cfg(target_os = "macos")]
-        assert_eq!(logs_dir, PathBuf::from("/Library/Logs/OpenFrame"));
-
-        #[cfg(target_os = "linux")]
-        assert_eq!(logs_dir, PathBuf::from("/var/log/openframe"));
-
-        #[cfg(target_os = "windows")]
-        {
-            let program_data = std::env::var_os("ProgramData").unwrap_or_default();
-            let expected = PathBuf::from(program_data).join("OpenFrame").join("logs");
-            assert_eq!(logs_dir, expected);
-        }
-    }
-
-    #[test]
-    fn test_get_app_support_directory() {
-        let app_dir = get_app_support_directory();
-
-        #[cfg(target_os = "macos")]
-        assert_eq!(
-            app_dir,
-            PathBuf::from("/Library/Application Support/OpenFrame")
-        );
-
-        #[cfg(target_os = "linux")]
-        assert_eq!(app_dir, PathBuf::from("/var/lib/openframe"));
-
-        #[cfg(target_os = "windows")]
-        {
-            let program_data = std::env::var_os("ProgramData").unwrap_or_default();
-            let expected = PathBuf::from(program_data).join("OpenFrame");
-            assert_eq!(app_dir, expected);
-        }
-    }
-
-    #[test]
-    fn test_get_secured_directory() {
-        let secured_dir = get_secured_directory();
-
-        #[cfg(target_os = "macos")]
-        assert_eq!(
-            secured_dir,
-            PathBuf::from("/Library/Application Support/OpenFrame/secured")
-        );
-
-        #[cfg(target_os = "linux")]
-        assert_eq!(secured_dir, PathBuf::from("/var/lib/openframe/secured"));
-
-        #[cfg(target_os = "windows")]
-        {
-            let program_data = std::env::var_os("ProgramData").unwrap_or_default();
-            let expected = PathBuf::from(program_data)
-                .join("OpenFrame")
-                .join("secured");
-            assert_eq!(secured_dir, expected);
-        }
-    }
-}
+#[path = "directories_tests.rs"]
+mod tests;
