+++ clients/openframe-client/src/platform/permissions.rs
@@ -1,7 +1,6 @@
 use std::fs::{self};
 use std::io;
 #[cfg(unix)]
-#[allow(unused_imports)] // MetadataExt used by linux-only paths
 use std::os::unix::fs::{MetadataExt, PermissionsExt};
 // Windows std::fs::Permissions already has readonly flag setters; no extra trait needed
 use std::path::Path;
@@ -22,17 +21,14 @@ static ADMIN_PRIVILEGES_GRANTED: AtomicBool = AtomicBool::new(false);
 
 /// Default UID for root user
 #[cfg(unix)]
-#[allow(dead_code)] // reserved for unix root checks; not currently referenced
 const ROOT_UID: u32 = 0;
 /// Default GID for admin group on macOS
 #[cfg(unix)]
 const ADMIN_GID: u32 = 80;
 
 #[cfg(not(unix))]
-#[allow(dead_code)] // parity with unix consts; not referenced on non-unix
 const ROOT_UID: u32 = 0;
 #[cfg(not(unix))]
-#[allow(dead_code)] // parity with unix consts; not referenced on non-unix
 const ADMIN_GID: u32 = 0;
 
 #[derive(Debug)]
@@ -98,8 +94,6 @@ impl Permissions {
                 let metadata = fs::metadata(path)?;
                 let mut perms = metadata.permissions();
                 #[cfg(target_os = "windows")]
-                #[allow(clippy::permissions_set_readonly_false)]
-                // clearing readonly is intended on windows
                 {
                     // Use cross-platform readonly flag instead of Windows-only bits
                     if perms.readonly() {
@@ -122,8 +116,6 @@ impl Permissions {
         }
 
         #[cfg(not(unix))]
-        #[allow(unreachable_code)]
-        // windows returns early above; tail is the fallback for other non-unix targets
         {
             // On Windows, we can check if the file is read-only if that's what we care about
             #[cfg(target_os = "windows")]
@@ -551,113 +543,5 @@ pub enum Capability {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use tempfile::tempdir;
-
-    #[test]
-    fn test_permissions_creation() {
-        let dir_perms = Permissions::directory();
-        assert_eq!(dir_perms.mode, 0o755);
-
-        let file_perms = Permissions::file();
-        assert_eq!(file_perms.mode, 0o644);
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn test_permissions_verification() {
-        if unsafe { libc::geteuid() } == 0 {
-            let temp = tempdir().unwrap();
-            let test_path = temp.path().join("test_file");
-            fs::write(&test_path, "test").unwrap();
-
-            let perms = Permissions::file();
-            assert!(perms.apply(&test_path).is_ok());
-            assert!(perms.verify(&test_path).unwrap());
-        }
-    }
-
-    #[test]
-    fn test_is_admin() {
-        // This just verifies the function runs without errors
-        let is_admin = PermissionUtils::is_admin();
-        println!("Running with admin privileges: {}", is_admin);
-    }
-
-    #[test]
-    fn test_has_capability() {
-        // Test all capabilities
-        for cap in &[
-            Capability::ManageServices,
-            Capability::WriteSystemDirectories,
-            Capability::ReadSystemLogs,
-            Capability::WriteSystemLogs,
-        ] {
-            let has_cap = PermissionUtils::has_capability(*cap);
-            println!("Has capability {:?}: {}", cap, has_cap);
-        }
-    }
-
-    #[test]
-    #[ignore = "triggers an interactive macOS osascript / Windows UAC admin prompt; hangs in headless CI"]
-    fn test_ensure_admin() {
-        // This should return Ok if already admin, or attempt to get privileges
-        let result = PermissionUtils::ensure_admin();
-
-        if PermissionUtils::is_admin() {
-            assert!(result.is_ok());
-        } else {
-            // The function might return Ok if the user granted privileges via the prompt,
-            // or an error if they declined or if there was an issue with the prompt
-            println!("Result of ensure_admin when not admin: {:?}", result);
-        }
-    }
-
-    #[test]
-    fn test_run_command() {
-        // Test running a simple command that should work on all platforms
-        // On Windows, use "cmd /c echo test"
-        // On Unix, use "echo test"
-        #[cfg(target_os = "windows")]
-        {
-            let result = PermissionUtils::run_command("cmd", &["/c", "echo", "test"]);
-            assert!(result.is_ok());
-        }
-
-        #[cfg(unix)]
-        {
-            let result = PermissionUtils::run_command("echo", &["test"]);
-            assert!(result.is_ok());
-        }
-    }
-
-    #[test]
-    fn test_cross_platform_permissions() {
-        // Create a temporary file and test platform-agnostic permissions
-        let temp = tempdir().unwrap();
-        let test_path = temp.path().join("test_file");
-        fs::write(&test_path, "test").unwrap();
-
-        // Test applying permissions
-        let perms = Permissions::file();
-        let result = perms.apply(&test_path);
-        assert!(result.is_ok());
-
-        // Test verifying permissions - should pass on all platforms
-        // even though the exact permission representation differs
-        let verify_result = perms.verify(&test_path);
-        assert!(verify_result.is_ok());
-
-        // Test retrieving permissions from a path
-        let retrieved_perms = Permissions::from_path(&test_path);
-        assert!(retrieved_perms.is_ok());
-    }
-
-    #[test]
-    fn test_can_read_system_logs() {
-        // Just verify the function runs without errors
-        let can_read = PermissionUtils::has_capability(Capability::ReadSystemLogs);
-        println!("Can read system logs: {}", can_read);
-    }
-}
+#[path = "permissions_tests.rs"]
+mod tests;
