+++ clients/openframe-client/src/platform/tool_updater/gui_app.rs
@@ -163,28 +163,5 @@ impl ToolUpdater for GuiAppToolUpdater {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::platform::DirectoryManager;
-    use std::path::Path;
-
-    #[test]
-    fn test_find_app_bundle_path() {
-        assert_eq!(
-            DirectoryManager::find_app_bundle_path(Path::new("/Applications/FAE Chat.app")),
-            Some(PathBuf::from("/Applications/FAE Chat.app"))
-        );
-
-        assert_eq!(
-            DirectoryManager::find_app_bundle_path(Path::new(
-                "/Applications/FAE Chat.app/Contents/MacOS/FAE Chat"
-            )),
-            Some(PathBuf::from("/Applications/FAE Chat.app"))
-        );
-
-        assert_eq!(
-            DirectoryManager::find_app_bundle_path(Path::new("/usr/bin/some-binary")),
-            None
-        );
-    }
-}
+#[path = "gui_app_tests.rs"]
+mod tests;
